+++ include/rain/algorithm.hpp
@@ -1,7 +1,11 @@
 // Uses min and max to avoid standardization issues.
 #pragma once
 
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
+
 #include <algorithm>
 
-using std::max;
 using std::min;
+using std::max;