+++ include/rain/networking/http/client.hpp
@@ -5,5 +5,4 @@
 
 namespace Rain::Networking::Http {
 	// TODO.
-	class Client : protected Http::Socket {};
 }