+++ include/rain/networking/node-service-host.hpp
@@ -7,15 +7,17 @@ namespace Rain::Networking {
 		public:
 		class Node {
 			public:
+			Node(const std::string &node) : node(node) {}
 			Node(const char *node) : node(node) {}
 
-			const char *getCStr() const noexcept { return this->node; }
+			const char *getCStr() const noexcept { return this->node.c_str(); }
 
 			private:
-			const char *node;
+			std::string node;
 		};
 		class Service {
 			public:
+			Service(const std::string &service) : service(service) {}
 			Service(const char *service) : service(service) {}
 			Service(std::size_t service) : service(std::to_string(service)) {}
 			Service(int service) : service(std::to_string(service)) {}
@@ -29,14 +31,7 @@ namespace Rain::Networking {
 		const Node node;
 		const Service service;
 
-		Host(const Node &node, const Service &service)
-				: node(node), service(service) {}
-		Host(const char *node, const char *service)
-				: node(node), service(service) {}
-		Host(const char *node, const std::string &service)
-				: node(node), service(service.c_str()) {}
-		Host(const char *node, std::size_t service)
-				: node(node), service(service) {}
-		Host(const char *node, int service) : node(node), service(service) {}
+		template <typename NodeType, typename ServiceType>
+		Host(NodeType node, ServiceType service) : node(node), service(service) {}
 	};
 }