+++ include/rain/networking/server.hpp
@@ -9,117 +9,77 @@
 #include <set>
 
 namespace Rain::Networking {
-	// Reference implementation.
-	template <typename ServerType, typename SlaveType, typename DataType>
-	class ServerSlave : virtual protected Socket {
+	// A new Slave is spawned from each accepted connection from the server.
+	class Slave : virtual protected Socket {
 		public:
-		// For use in accessing subclass type from superclass.
-		typedef ServerType Server;
-
-		Server *server;
-		DataType data;
-
-		ServerSlave(Socket &socket, Server *server)
-				: Socket(std::move(socket)), server(server) {}
+		Slave(Socket &socket) : Socket(std::move(socket)) {}
 
 		// Expose some relevant functions from base Socket.
 		using Socket::getNativeSocket;
 		using Socket::getFamily;
-		using Socket::getType;
 		using Socket::getProtocol;
+		using Socket::getType;
+		using Socket::isValid;
 		using Socket::send;
 		using Socket::recv;
 		using Socket::shutdown;
 		using Socket::close;
 	};
 
-	// Handles listening and accepting of new connections, as well as creating the
-	// HttpServerSlave to be configured correctly.
+	// DataType is data carried with each new connected slave socket.
 	template <typename SlaveType>
 	class Server : virtual protected Socket {
+		// Shorthand for use in subclasses defining virtual functions.
 		public:
-		// Shorthand.
 		typedef SlaveType Slave;
-		// If server is closed, can take up to this much time for thread to stop.
-		std::chrono::milliseconds acceptTimeoutMs{1000};
 
-		// Subclasses should override this function with their `this` pointer.
+		// Subclass behavior.
 		protected:
-		virtual void *getSubclassPtr() { return reinterpret_cast<void *>(this); }
+		// Factory: create new slave from accepted Socket.
+		virtual Slave *onServerAccept(Socket &acceptedSocket) {
+			return new Slave(acceptedSocket);
+		}
 
-		// Handlers.
-		virtual void onSlaveTask(SlaveType *) {}
+		// Handler: when new slave is created and ready.
+		virtual void onBeginSlaveTask(Slave &) noexcept = 0;
 
-		// Slaves management.
-		ThreadPool threadPool;
+		// Settings.
+		private:
+		// Slave management.
+		ThreadPool<Slave *> threadPool;
 
-		// Keeps track of active slaves.
-		std::set<SlaveType *> slaves;
+		// Active slaves.
+		std::set<Slave *> slaves;
 		std::mutex slavesMtx;
 
-		// Constructor.
+		// If server is closed, can take up to this much time for thread to stop.
+		const std::chrono::milliseconds ACCEPT_TIMEOUT_MS;
+
+		// Getters.
 		public:
-		Server(std::size_t maxThreads = 0) : Socket(), threadPool(maxThreads) {}
+		using Socket::getFamily;
+		using Socket::getNativeSocket;
+		using Socket::getProtocol;
+		using Socket::getType;
+		using Socket::isValid;
+		using Socket::getService;
+
+		// Constructor.
+		Server(std::size_t maxThreads = 1024,
+			const std::chrono::milliseconds &ACCEPT_TIMEOUT_MS =
+				std::chrono::milliseconds(1000))
+				: Socket(),
+					threadPool(maxThreads),
+					ACCEPT_TIMEOUT_MS(ACCEPT_TIMEOUT_MS) {}
 		~Server() {
+			// Close can always throw exception if threads execute in weird order.
 			try {
 				this->close();
 			} catch (...) {
 			}
 		}
 
-		// Bind, listen, and accept continuously until master is closed.
-		void serve(const Host &host,
-			bool blocking = true,
-			int backlog = SOMAXCONN) {
-			this->bind(host);
-			this->listen(backlog);
-
-			// This function needs to exit immediately once the server has been
-			// destructed, since its member variables will no longer be valid.
-			std::function<void(void *)> serveSync = [this](void *param) {
-				// Continuously accept new connections until master is closed.
-				while (true) {
-					SlaveType *slave = NULL;
-					try {
-						while (slave == NULL) {
-							Socket acceptedSocket(
-								this->accept(NULL, NULL, this->acceptTimeoutMs));
-							if (acceptedSocket.getNativeSocket() != NATIVE_SOCKET_INVALID) {
-								slave = new SlaveType(acceptedSocket,
-									reinterpret_cast<typename SlaveType::Server *>(
-										this->getSubclassPtr()));
-							}
-						}
-					} catch (...) {
-						// Accept exception, server has likely shut down.
-						break;
-					}
-					this->slavesMtx.lock();
-					this->slaves.insert(slave);
-					this->slavesMtx.unlock();
-
-					// For each new connection, spawn a task to handle it.
-					this->threadPool.queueTask(
-						[this](void *vpSlave) {
-							SlaveType *slave = reinterpret_cast<SlaveType *>(vpSlave);
-							this->onSlaveTask(slave);
-							this->slavesMtx.lock();
-							this->slaves.erase(slave);
-							delete slave;
-							this->slavesMtx.unlock();
-						},
-						reinterpret_cast<void *>(slave));
-				}
-			};
-
-			if (blocking) {
-				serveSync(reinterpret_cast<void *>(this));
-			} else {
-				this->threadPool.queueTask(serveSync, reinterpret_cast<void *>(this));
-			}
-		}
-
-		// Shutting down the server should shutdown all associated slaves as well.
+		// Close server and slaves and wait for tasks to complete.
 		void close() {
 			Socket::close();
 
@@ -133,20 +93,51 @@ namespace Rain::Networking {
 			this->threadPool.blockForTasks();
 		}
 
-		// Getters.
-		Host::Service getService() {
-			struct sockaddr_in sin;
-			socklen_t addrlen = sizeof(sin);
-			getsockname(this->getNativeSocket(),
-				reinterpret_cast<struct sockaddr *>(&sin),
-				&addrlen);
-			return Host::Service(static_cast<std::size_t>(ntohs(sin.sin_port)));
-		}
+		public:
+		// Bind, listen, and accept continuously until master is closed.
+		void serve(const Host &host,
+			bool blocking = true,
+			int backlog = SOMAXCONN) {
+			this->bind(host);
+			this->listen(backlog);
+			this->threadPool.queueTask(
+				[this](Slave *) {
+					// This function needs to exit immediately once the server has been
+					// destructed, since its member variables will no longer be valid.
+					while (true) {
+						try {
+							Socket acceptedSocket(
+								this->accept(NULL, NULL, this->ACCEPT_TIMEOUT_MS));
+							if (acceptedSocket.isValid()) {
+								this->slavesMtx.lock();
+								Slave *slave =
+									*this->slaves.insert(this->onServerAccept(acceptedSocket))
+										 .first;
+								this->slavesMtx.unlock();
 
-		// Expose some additional getter functions from Socket.
-		using Socket::getNativeSocket;
-		using Socket::getFamily;
-		using Socket::getType;
-		using Socket::getProtocol;
+								// For each new connection, spawn a task to handle it.
+								this->threadPool.queueTask(
+									[this](Slave *slave) {
+										this->onBeginSlaveTask(*slave);
+										this->slavesMtx.lock();
+										this->slaves.erase(slave);
+										delete slave;
+										this->slavesMtx.unlock();
+									},
+									slave);
+							}
+						} catch (...) {
+							// Accept exception, server has likely shut down.
+							break;
+						}
+					}
+				},
+				NULL);
+
+			// Blocking servers need to be closed in another thread.
+			if (blocking) {
+				this->threadPool.blockForTasks();
+			}
+		}
 	};
 }